@@ -1,49 +1,50 @@
-const express = require("express");
-const morgan = require("morgan");
-const mongoose = require("mongoose");
+const express = require('express');
+const morgan = require('morgan');
+const connectDB = require('./config/db');
+const path = require('path');
 
 // Middleware
-const cors = require("cors");
+const cors = require('cors');
 
-require("dotenv").config();
-const logs = require("./api/logs");
+require('dotenv').config();
+
+const logsRouter = require('./routes/logRoutes');
 
 const app = express();
 
 // Body parser
 app.use(express.json());
 
-// MongoDB Connection
-mongoose
-  .connect(process.env.ATLAS_URI, {
-    useNewUrlParser: true,
-    useCreateIndex: true,
-    useUnifiedTopology: true,
-  })
-  .catch((err) => console.log(err));
-
-mongoose.connection.once("open", () => {
-  console.log("mongoDB connection successful");
-});
-
-app.use(morgan("common"));
+app.use(morgan('common'));
 app.use(cors());
 
-app.use("/api/logs", logs);
+app.use('/api/listLogEntries', logsRouter);
+app.use('/api/postLogEntry', logsRouter);
+app.use('/api/editLogEntry', logsRouter);
+app.use('/api/deleteLogEntry', logsRouter);
 
 // deployment
-if (process.env.NODE_ENV === "production") {
-  // set static folder
-  const path = require("path");
-  app.use(express.static("client/build"));
+if (process.env.NODE_ENV === 'production') {
+  const __dirname = path.resolve();
+  app.use(express.static(path.join(__dirname, './client/build')));
 
-  app.get("*", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
+  app.get('*', (req, res) => {
+    res.sendFile(
+      path.join(__dirname, './client/build/index.html'),
+      function (err) {
+        res.status(500).send(err);
+      }
+    );
   });
+} else {
+  app.get('/', (req, res) => res.send('server is ready'));
 }
-
 // Server port
 const PORT = process.env.PORT || 8080;
-app.listen(PORT, () => {
-  console.log(`listening at http://localhost:${PORT}`);
+
+// Connecting MongoDB Atlas then server
+connectDB().then(() => {
+  app.listen(PORT, () => {
+    console.log(`listening to the Server at http://localhost:${PORT}`);
+  });
 });
